backend/src/requirements.txt
pymongo
python-dotenv
pyjwt
werkzeug
